E:40472 [*] Small logic changes

+++ src/classes/XLite/Model/Base/Image.php
@@ -86,9 +86,9 @@ abstract class Image extends \XLite\Model\Base\Storage
      * @see   ____var_see____
      * @since 1.0.0
      *
-     * @Column (type="fixedstring", length="32")
+     * @Column (type="fixedstring", length="32", nullable=true)
      */
-    protected $hash = '';
+    protected $hash;
 
     /**
      * Get image URL for customer front-end
@@ -177,15 +177,21 @@ abstract class Image extends \XLite\Model\Base\Storage
      */
     public function checkImageHash()
     {
-        list($path, $isTempFile) = $this->getLocalPath();
+        $result = true;
 
-        $hash = \Includes\Utils\FileManager::getHash($path);
+        if ($this->getHash()) {
+            list($path, $isTempFile) = $this->getLocalPath();
 
-        if ($isTempFile) {
-            \Includes\Utils\FileManager::deleteFile($path);
+            $hash = \Includes\Utils\FileManager::getHash($path);
+
+            if ($isTempFile) {
+                \Includes\Utils\FileManager::deleteFile($path);
+            }
+
+            $result = $this->getHash() === $hash;
         }
 
-        return $this->getHash() === $hash;
+        return $result;
     }
 
     /**
@@ -247,10 +253,13 @@ abstract class Image extends \XLite\Model\Base\Storage
 
             if (is_array($data)) {
 
-                $this->width    = $data[0];
-                $this->height   = $data[1];
-                $this->mime     = $data['mime'];
-                $this->hash     = \Includes\Utils\FileManager::getHash($path);
+                $this->setWidth($data[0]);
+                $this->setHeight($data[1]);
+                $this->setMime($data['mime']);
+                $hash = \Includes\Utils\FileManager::getHash($path);
+                if ($hash) {
+                    $this->setHash($hash);
+                }
 
             } else {
                 $result = false;
